Refactor helper functions to prevent dependency loop
