+++ packages/shared/src/test_utils/common.js
@@ -1,10 +1,10 @@
 const expect = require('chai').expect;
 const jsdom = require('jsdom');
-const Url = require('../utils/url/url.js');
+const reset = require('../utils/url/url').reset;
 
 const setURL = url => {
     jsdom.changeURL(window, url);
-    Url.reset();
+    reset();
 };
 
 module.exports = {
