+++ packages/shared/src/utils/storage/index.js
@@ -1,2 +1,1 @@
-export * from './storage_events';
 export * from './storage';
